Wrap zle-line-init to support fetching suggestions on line init
Use case suggested by @romkatv uses zle-line-init to restore buffer
after running a widget to cd up one level (GitHub #431).

As far as I can tell, the ignoring of zle-line-* was added in commit
9788c2e to support some deprecation warnings that were removed some time
ago.

The pattern was then widened in commit 0c940e7 to zle-* to fix problems
encountered when wrapping zle-isearch-update.

This commit removes the hard coded ignore of all zle-* widgets and puts
zle-isearch-update and zle-keymap-select into the default list of
widgets to be ignored. It leaves zle-line-init and zle-line-finish not
ignored so that suggestions will be fetched after they are run.